local-services/ingestion-service/requirements.txt
fastapi[standard]
google-cloud-pubsub